% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{bar_plot}
\alias{bar_plot}
\title{Graph bar_plot}
\usage{
bar_plot(df, x_variable, y_variable, bar_color = "black", main_title,
  type, color2)
}
\arguments{
\item{radian}{@name bar_plot @description easy way to bar plot}
}
\value{
value
}
\description{
Graph bar_plot
}
