% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{box_plot}
\alias{box_plot}
\title{Graph box_plot}
\usage{
box_plot(df, x_variable, y_variable, main_title)
}
\arguments{
\item{radian}{@name box_plot @description easy way to box plot}
}
\value{
value
}
\description{
Graph box_plot
}
