% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram_plot}
\alias{histogram_plot}
\title{Graph Histograms}
\usage{
histogram_plot(df, x_variable, outline_color = "black",
  bin_color = "white", main_title)
}
\arguments{
\item{radian}{@name histogram_plot @description provides an easy way to view histograms}
}
\value{
value
}
\description{
Graph Histograms
}
