% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalplot.R
\name{normal_plots_check}
\alias{normal_plots_check}
\title{Graph Normal Distributions}
\usage{
normal_plots_check(vec, data_name, measure_name, color_desired)
}
\arguments{
\item{radian}{@name normal_plots_check @description provides an easy way to view if the data is normal}
}
\value{
value
}
\description{
Graph Normal Distributions
}
