% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_anova_1.R
\name{summarize_anova}
\alias{summarize_anova}
\title{summarize_anova}
\usage{
summarize_anova(df, x_variable, y_variable)
}
\arguments{
\item{radian}{@name ss anova @description}
}
\value{
value
}
\description{
summarize_anova
}
