% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_one_qual_variable.R
\name{summarize_one_qual_variable}
\alias{summarize_one_qual_variable}
\title{summarize_1_qualitative}
\usage{
summarize_one_qual_variable(df, x_variable)
}
\arguments{
\item{radian}{@name ss 1 qualitative @description}
}
\value{
value
}
\description{
summarize_1_qualitative
}
