% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_one_quant_variable.R
\name{summarize_one_quant_variable}
\alias{summarize_one_quant_variable}
\title{summarize_1_quantitative}
\usage{
summarize_one_quant_variable(df, x_variable)
}
\arguments{
\item{radian}{@name ss 1 quantitative @description}
}
\value{
value
}
\description{
summarize_1_quantitative
}
