% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{summarize_one_variable}
\alias{summarize_one_variable}
\title{Graph plot}
\usage{
summarize_one_variable(df, x_variable)
}
\arguments{
\item{radian}{@name ss @description}
}
\value{
value
}
\description{
Graph plot
}
