% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_slr.R
\name{summarize_slr}
\alias{summarize_slr}
\title{summarize_slr}
\usage{
summarize_slr(df, x_variable, y_variable)
}
\arguments{
\item{radian}{@name ss slr @description}
}
\value{
value
}
\description{
summarize_slr
}
