% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_slr_.R
\name{summarize_slr_}
\alias{summarize_slr_}
\title{summarize_slr_}
\usage{
summarize_slr_(df, x_variable, y_variable)
}
\arguments{
\item{radian}{@name ss slr @description}
}
\value{
value
}
\description{
summarize_slr_
}
