% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_two_qual_variable.R
\name{summarize_two_qual_variable}
\alias{summarize_two_qual_variable}
\title{summarize_2_qualitative}
\usage{
summarize_two_qual_variable(df, x_variable, y_variable)
}
\arguments{
\item{radian}{@name ss 2 qualitative @description}
}
\value{
value
}
\description{
summarize_2_qualitative
}
