% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_by_group.R
\name{summary_by_group}
\alias{summary_by_group}
\title{summary_by_group}
\usage{
summary_by_group(df, x_variable, y_variable)
}
\arguments{
\item{radian}{@name ss after_anova @description}
}
\value{
value
}
\description{
summary_by_group
}
