% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesanalysis.R
\name{time_series_analysis}
\alias{time_series_analysis}
\title{Graph time_series_plot}
\usage{
time_series_analysis(df, y_variable, smooth_technique, start_date,
  end_date, by_n = 1, main_title, w = 0.2, L = 5)
}
\arguments{
\item{radian}{@name scatter_plot @description provides an easy way to view if data is norma;}
}
\value{
value
}
\description{
Graph time_series_plot
}
