% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesplot.R
\name{time_series_plot}
\alias{time_series_plot}
\title{Graph scatter_plot}
\usage{
time_series_plot(df, y_variable, line_color = "red", main_title)
}
\arguments{
\item{radian}{@name scatter_plot @description provides an easy way to view if data is norma;}
}
\value{
value
}
\description{
Graph scatter_plot
}
